% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{assign_cd8_labels}
\alias{assign_cd8_labels}
\title{Assign CD8+ T cell Labels}
\usage{
assign_cd8_labels(
  seurat_obj,
  all_markers,
  naive_markers,
  cytotoxic_markers,
  exhausted_markers
)
}
\arguments{
\item{seurat_obj}{A Seurat object with CD8+ T cell expression data.}

\item{all_markers}{A data frame of all markers identified across clusters.}

\item{naive_markers}{A vector of naive CD8+ T cell markers.}

\item{cytotoxic_markers}{A vector of cytotoxic CD8+ T cell markers.}

\item{exhausted_markers}{A vector of exhausted CD8+ T cell markers.}
}
\value{
A list containing the updated Seurat object with CD8+ labels and a subset of CD8+ T cells.
}
\description{
This function assigns labels to CD8+ T cells based on marker gene expression (e.g., naive, cytotoxic, exhausted).
}
