% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{calculate_DEGs}
\alias{calculate_DEGs}
\title{Perform Differential Expression Analysis}
\usage{
calculate_DEGs(rna_data, condition_column, alpha, log2FC_threshold)
}
\arguments{
\item{rna_data}{A \code{DESeqDataSet} object containing the RNA-Seq count matrix and metadata.}

\item{condition_column}{A string specifying the name of the column in the metadata
that defines the experimental conditions/groups.}

\item{alpha}{A numeric value specifying the significance threshold for the p-value (e.g., 0.05).}

\item{log2FC_threshold}{A numeric value specifying the log2 fold change threshold to
determine upregulated or downregulated genes.}
}
\value{
A data frame containing the DESeq2 results with an additional column, \code{significance},
indicating whether each gene is upregulated, downregulated, or not significant.
}
\description{
This function calculates differentially expressed genes (DEGs) from RNA-Seq data
using the DESeq2 package. It applies thresholds for statistical significance (p-value)
and fold change to classify genes as upregulated, downregulated, or not significant.
}
