% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{cluster_and_umap}
\alias{cluster_and_umap}
\title{Cluster and Perform UMAP}
\usage{
cluster_and_umap(seurat_obj, dims = 20, resolution = 0.1)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA-seq data.}

\item{dims}{Number of dimensions to use for clustering (default is 20).}

\item{resolution}{Resolution parameter for clustering (default is 0.1).}
}
\value{
A Seurat object with clustering and UMAP results added.
}
\description{
This function clusters the cells and performs UMAP dimensionality reduction for visualization.
}
