% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{extract_genes_hr_gt1}
\alias{extract_genes_hr_gt1}
\title{Extract Significant Genes with Hazard Ratio Greater Than 1 and Save to CSV}
\usage{
extract_genes_hr_gt1(results, output_file = "significant_genes.csv")
}
\arguments{
\item{results}{A DataFrame containing gene analysis results with at least
the following columns: \code{significant}, \code{hazard_ratio},
and \code{gene_name}.}

\item{output_file}{Character. The name of the output \code{.csv} file to save
the results. Default is "significant_genes.csv".}
}
\value{
A DataFrame with two columns:
\itemize{
\item \code{gene_name}: Names of significant genes with HR > 1.
\item \code{hazard_ratio}: The hazard ratios of the significant genes.
}
}
\description{
This function filters significant genes with a hazard ratio (HR) greater than 1
from a given results DataFrame. It returns a DataFrame of the filtered genes
and their hazard ratios and saves the results to a \code{.csv} file.
}
