% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{filter_rna_data}
\alias{filter_rna_data}
\title{Filter RNA Data}
\usage{
filter_rna_data(
  rna_data,
  condition_column,
  condition_levels,
  na_threshold = 0.8
)
}
\arguments{
\item{rna_data}{A \code{SummarizedExperiment} object containing RNA-Seq count data.}

\item{condition_column}{A character string specifying the column name in \code{colData}
that contains the conditions for filtering.}

\item{condition_levels}{A character vector of condition levels to retain in the data.}

\item{na_threshold}{A numeric value (default = 0.8) specifying the maximum allowed
fraction of NAs in a row. Rows with a higher fraction of NAs will be removed.}
}
\value{
A filtered \code{SummarizedExperiment} object with updated counts, \code{rowRanges},
and \code{colData}.
}
\description{
This function filters RNA-Seq data stored in a \code{SummarizedExperiment} object.
It allows for filtering samples based on specific conditions, removing rows with
high NA values, replacing remaining NAs with the median, and removing low-count rows.
}
