% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{filter_seurat_object}
\alias{filter_seurat_object}
\title{Filter Seurat Object Based on Quality Control Criteria}
\usage{
filter_seurat_object(
  seurat_obj,
  min_features = 200,
  max_features = 2500,
  max_percent_mt = 5
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA-seq data.}

\item{min_features}{An integer specifying the minimum number of detected features (genes) required for a cell to be retained. Default is 200.}

\item{max_features}{An integer specifying the maximum number of detected features (genes) allowed for a cell. Default is 2500.}

\item{max_percent_mt}{A numeric value specifying the maximum percentage of mitochondrial genes allowed for a cell. Default is 5.}
}
\value{
A filtered Seurat object, where cells meeting the quality control criteria are retained.
}
\description{
This function filters a Seurat object based on the number of detected features (genes) and the percentage of mitochondrial genes.
It ensures that only high-quality cells are kept for downstream analysis.
}
\details{
The function adds a new metadata column to the Seurat object, \code{percent.mt}, which represents the percentage of mitochondrial genes expressed in each cell. The cells are then filtered based on the following criteria:
\itemize{
\item The number of detected features (genes) must be between \code{min_features} and \code{max_features}.
\item The percentage of mitochondrial gene expression must be below \code{max_percent_mt}.
}

This helps to remove cells with poor quality (e.g., low gene count or high mitochondrial gene expression) that could skew downstream analyses.
}
