% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{get_mutated_samples}
\alias{get_mutated_samples}
\title{Get Unique Mutated Sample IDs for Selected Genes}
\usage{
get_mutated_samples(cancer_maf, genes_of_interest)
}
\arguments{
\item{cancer_maf}{A MAF object created using the \code{maftools} package, containing mutation data.}

\item{genes_of_interest}{A character vector of gene names for which the mutated samples are to be extracted.}
}
\value{
A character vector containing the unique tumor sample barcodes of samples where mutations in the selected genes are present.
}
\description{
This function extracts unique tumor sample barcodes for selected genes with mutations from a MAF (Mutation Annotation Format) dataset.
}
\details{
\itemize{
\item The function subsets the MAF data to include only the mutations for the specified genes.
\item It then extracts the unique tumor sample barcodes where mutations for those genes are present.
}
}
