% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{get_top_combined_genes}
\alias{get_top_combined_genes}
\title{Get Top Combined Ranked Genes and Save to CSV}
\usage{
get_top_combined_genes(
  graph_test_res,
  q_value_threshold = 0.05,
  top_n = 10,
  output_file = "top_genes.csv"
)
}
\arguments{
\item{graph_test_res}{A data frame containing gene statistics,
including \code{q_value} (adjusted p-value) and \code{morans_I} (Moran's I statistic).}

\item{q_value_threshold}{Numeric. The threshold for significant q-values.
Default is 0.05.}

\item{top_n}{Integer. The number of top genes to return based on the
combined rank. Default is 10.}

\item{output_file}{Character. The name of the output \code{.csv} file to save
the results. Default is "top_genes.csv".}
}
\value{
A character vector containing the names of the top-ranked genes.
}
\description{
This function identifies the top genes based on a combined rank of
q-values and Moran's I statistic from spatial transcriptomics analysis,
and saves the results to a \code{.csv} file.
}
\details{
The function filters genes based on the \code{q_value_threshold}, then combines
the ranks of \code{q_value} (ascending order) and \code{morans_I} (descending order).
The combined rank is used to determine the top \code{top_n} genes.
The top genes and their ranks are saved to a \code{.csv} file.
}
