% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{mapping_genes}
\alias{mapping_genes}
\title{Map Differentially Expressed Genes (DEGs) with Gene Metadata}
\usage{
mapping_genes(deg_results, gene_metadata_dt, logFC_threshold, padj_threshold)
}
\arguments{
\item{deg_results}{A data frame containing differential expression results. The data frame should include at least the following columns:
\describe{
\item{log2FoldChange}{The log2 fold change values for the genes.}
\item{padj}{The adjusted p-values for the genes.}
\item{ensemble_id}{A unique identifier for each gene (can be rownames if not present as a column).}
}}

\item{gene_metadata_dt}{A data frame containing gene metadata. This should include at least the following columns:
\describe{
\item{ensemble_id}{The unique identifier for the gene, which matches the \code{ensemble_id} in \code{deg_results}.}
\item{gene_name}{The gene names corresponding to the \code{ensemble_id}.}
}}

\item{logFC_threshold}{A numeric value specifying the threshold for log2 fold change. Genes with \code{log2FoldChange > logFC_threshold} are considered upregulated,
and genes with \code{log2FoldChange < -logFC_threshold} are considered downregulated.}

\item{padj_threshold}{A numeric value specifying the adjusted p-value threshold. Only genes with \code{padj < padj_threshold} will be selected.}
}
\value{
A data frame containing the upregulated and downregulated genes based on the given thresholds, with columns:
\describe{
\item{ensemble_id}{The unique identifier for the gene.}
\item{gene_name}{The gene name associated with the \code{ensemble_id}.}
\item{log2FoldChange}{The log2 fold change values for the gene.}
\item{padj}{The adjusted p-values for the gene.}
}
}
\description{
This function takes in the results of differential gene expression analysis (\code{deg_results}), a gene metadata table (\code{gene_metadata_dt}),
and filters for upregulated and downregulated genes based on user-defined thresholds for log fold change and adjusted p-value.
It then returns a merged table with gene names and the associated differential expression data.
}
