% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{perform_pca}
\alias{perform_pca}
\title{Perform Principal Component Analysis (PCA) on Seurat Object}
\usage{
perform_pca(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA-seq data.}
}
\value{
A Seurat object with PCA results stored in the \code{pca} assay, ready for further analysis (e.g., visualization, clustering).
}
\description{
This function performs PCA on a Seurat object after normalizing the data, identifying variable features, and scaling the data.
The resulting PCA components can be used for dimensionality reduction and visualization in downstream analysis.
}
\details{
The function performs the following steps:
\itemize{
\item Normalizes the data using \code{Seurat::NormalizeData()}.
\item Identifies highly variable features with \code{Seurat::FindVariableFeatures()}.
\item Scales the data using \code{Seurat::ScaleData()}.
\item Runs PCA using the identified variable features with \code{Seurat::RunPCA()}.
}
These steps are essential for reducing the dimensionality of the data and identifying principal components that explain most of the variance in the dataset.
}
