% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{perform_pseudotime}
\alias{perform_pseudotime}
\title{Perform Pseudotime Analysis}
\usage{
perform_pseudotime(seurat_obj, num_dim = 100, root_cluster = "Naive CD8+ T")
}
\arguments{
\item{seurat_obj}{A Seurat object with CD8+ T cells.}

\item{num_dim}{Number of dimensions for preprocessing (default is 100).}

\item{root_cluster}{The cluster to set as the root (default is "Naive CD8+ T").}
}
\value{
A Monocle3 CellDataSet with pseudotime trajectory learned.
}
\description{
This function performs pseudotime analysis on CD8+ T cells using the Monocle3 package.
}
