% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{perform_survival_analysis}
\alias{perform_survival_analysis}
\title{Perform Survival Analysis on RNA-Seq Data}
\usage{
perform_survival_analysis(
  count_matrix_filtered,
  gene_metadata_dt,
  clinical_data_filtered
)
}
\arguments{
\item{count_matrix_filtered}{A matrix of RNA-Seq count data, where rows represent genes and columns represent samples.}

\item{gene_metadata_dt}{A data frame containing metadata for the genes, including a \code{gene_name} column.}

\item{clinical_data_filtered}{A data frame containing clinical data, including \code{submitter_id}, \code{overall_survival}, and \code{deceased} columns.}
}
\value{
A data frame with the results of the survival analysis for each gene, including:
\itemize{
\item \code{gene_name}: The name of the gene.
\item \code{hazard_ratio}: The hazard ratio for the gene.
\item \code{pvalue}: The p-value from the Cox regression for the gene.
\item \code{significant}: Whether the gene is significantly associated with survival (p-value ≤ 0.05).
}
}
\description{
This function performs survival analysis on RNA-Seq data by grouping samples into two strata (HIGH/LOW) based on the median expression of each gene. It then performs Cox regression analysis for each gene and returns the hazard ratios and p-values.
}
