% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{plot_cd8_trajectory}
\alias{plot_cd8_trajectory}
\title{Visualize Pseudotime Trajectory}
\usage{
plot_cd8_trajectory(
  cds,
  cd8_tcells,
  color_by_pseudotime = TRUE,
  label_size = 6
)
}
\arguments{
\item{cds}{A Monocle3 CellDataSet containing pseudotime data.}

\item{cd8_tcells}{A Seurat object containing CD8+ T cell clusters.}

\item{color_by_pseudotime}{A logical value to color by pseudotime (default is TRUE).}

\item{label_size}{Size of labels in the plot (default is 6).}
}
\value{
Pseudotime trajectory plot.
}
\description{
This function visualizes the pseudotime trajectory for CD8+ T cells.
}
