% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{plot_km_curves}
\alias{plot_km_curves}
\title{Plot Kaplan-Meier Curves for Significant Genes}
\usage{
plot_km_curves(strata_data, clinical_data_filtered, significant_genes)
}
\arguments{
\item{strata_data}{A data frame containing gene expression data for multiple cases. This data frame should include at least the columns:
\describe{
\item{gene_name}{The name of the gene.}
\item{case_id}{The identifier for the case/sample.}
}}

\item{clinical_data_filtered}{A data frame containing clinical data filtered for the relevant cases. This data frame should include at least the columns:
\describe{
\item{submitter_id}{The identifier for the case/sample.}
\item{overall_survival}{The survival time of the patient (in days).}
\item{deceased}{A binary indicator of whether the patient is deceased (1 = deceased, 0 = alive).}
\item{strata}{A factor or categorical variable used to stratify the survival analysis (e.g., treatment groups).}
}}

\item{significant_genes}{A character vector containing the names of significant genes to plot. The function will generate Kaplan-Meier curves for each gene in this vector.}
}
\value{
A list of Kaplan-Meier plots (\code{ggsurvplot} objects) for each significant gene. The plots are printed individually for each gene.
}
\description{
This function generates Kaplan-Meier survival curves for each significant gene based on its association with the clinical data.
The function filters the \code{strata_data} for the significant genes, merges it with clinical data, and fits the Kaplan-Meier model
for each gene. A survival plot is then created for each gene showing survival probability over time.
}
