% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{plot_mutation_counts}
\alias{plot_mutation_counts}
\title{Plot Mutation Counts for Each Gene}
\usage{
plot_mutation_counts(mutation_counts)
}
\arguments{
\item{mutation_counts}{A data.frame or data.table containing the mutation data, including columns for:
\itemize{
\item \code{Gene}: The gene name.
\item \code{Mutation_Type}: The type of mutation (e.g., missense, nonsense).
\item \code{NumSamples}: The number of samples with that mutation type for the given gene.
}}
}
\value{
A series of bar plots, one for each gene, showing the count of mutations by mutation type.
}
\description{
This function generates separate bar plots showing the mutation counts for each gene, based on mutation types and sample counts.
}
\details{
The function iterates over each unique gene in the \code{mutation_counts} data and creates a bar plot for that gene. Each plot represents the count of mutations per mutation type for the gene, with the bars filled according to the mutation type.
}
