% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{plot_results}
\alias{plot_results}
\title{Plot Survival Analysis Results}
\usage{
plot_results(results)
}
\arguments{
\item{results}{A data frame containing the results of the survival analysis. The data frame should have the following columns:
\describe{
\item{gene_name}{Character string, the name of the gene.}
\item{hazard_ratio}{Numerical value representing the hazard ratio.}
\item{pvalue}{Numerical value representing the p-value of the gene.}
\item{significant}{Character string, "Yes" if the gene is statistically significant, otherwise "No".}
}}
}
\value{
A list containing three ggplot objects:
\describe{
\item{SignificancePlot}{A bar plot showing the count of significant vs. non-significant genes.}
\item{HRPlot}{A bar plot showing the count of significant genes categorized by hazard ratio.}
\item{VolcanoPlot}{A volcano plot showing the relationship between log hazard ratio and -log10 p-value.}
}
}
\description{
This function generates three types of plots to visualize the survival analysis results:
a bar plot showing the number of significant vs. non-significant genes,
a bar plot showing the count of significant genes categorized by hazard ratio,
and a volcano plot representing the relationship between hazard ratio and p-value.
}
