% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess RNA-Seq and Clinical Data}
\usage{
preprocess_data(rna_data, clinical_data, extracted_mutated_sample_ids)
}
\arguments{
\item{rna_data}{A \code{SummarizedExperiment} or similar object containing RNA-Seq count data. It must have a count matrix and associated gene metadata.}

\item{clinical_data}{A data frame containing clinical data with a column \code{submitter_id} that matches sample IDs in \code{rna_data}.}

\item{extracted_mutated_sample_ids}{A character vector of sample IDs that have mutations of interest.}
}
\value{
A list containing two elements:
\itemize{
\item \code{count_matrix_filtered}: A matrix of RNA-Seq count data for samples that have mutation data and are present in both datasets.
\item \code{clinical_data_filtered}: A data frame of clinical data for samples that have mutation data and are present in both datasets.
}
}
\description{
This function preprocesses RNA-Seq count data and clinical data, filtering both to retain only the samples that have mutation data and are present in both the RNA-Seq and clinical datasets. It also standardizes the sample IDs to a common format.
}
\details{
The function filters the RNA-Seq count matrix and the clinical data to include only the samples that are present in both the mutation data and clinical data. The sample IDs are standardized to the first three components (using hyphen delimiters) for consistency.
}
