% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{run_graph_test}
\alias{run_graph_test}
\title{Run Graph Test}
\usage{
run_graph_test(cds, neighbor_graph = "knn", cores = 1)
}
\arguments{
\item{cds}{A Monocle3 CellDataSet to run the graph test on.}

\item{neighbor_graph}{The neighbor graph type to use ("knn" or "principal_graph").}

\item{cores}{Number of CPU cores to use for parallel processing.}
}
\value{
Data frame with graph test results.
}
\description{
This function performs a graph test on the Monocle3 CellDataSet.
}
