% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{standardize_tcga_ids}
\alias{standardize_tcga_ids}
\title{Standardize TCGA Sample IDs}
\usage{
standardize_tcga_ids(sample_ids)
}
\arguments{
\item{sample_ids}{A character vector of TCGA sample IDs.}
}
\value{
A character vector of standardized TCGA sample IDs, where each ID is truncated to the first three components, or \code{NA} if the format is invalid.
}
\description{
This function standardizes TCGA sample IDs by extracting the first three components of the ID before the first hyphen ("-"). If the ID format is invalid (does not contain "-"), it returns \code{NA}.
}
\details{
The function ensures that all TCGA sample IDs are standardized to the format containing only the first three components separated by hyphens. If the input contains IDs in an incorrect format (i.e., not containing "-"), the function returns \code{NA} for those IDs.
}
