% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{style_mutation_table}
\alias{style_mutation_table}
\title{Style the Mutation Counts Table}
\usage{
style_mutation_table(mutation_counts)
}
\arguments{
\item{mutation_counts}{A data frame or data table containing the mutation counts, with columns for:
\itemize{
\item \code{Gene}: The gene name.
\item \code{Mutation_Type}: The type of mutation.
\item \code{NumSamples}: The number of samples with that mutation type for the given gene.
}}
}
\value{
A styled table with mutation counts for selected genes, formatted for better presentation.
}
\description{
This function applies styling to a mutation counts table, enhancing the readability and presentation of the data using the \code{kable} and \code{kableExtra} packages.
}
\details{
The function uses \code{kableExtra::kbl()} to create a table and then applies various styling options:
\itemize{
\item \code{striped}, \code{hover}, \code{condensed}, and \code{responsive} bootstrap options for table formatting.
\item The first column (Gene) is made bold with a fixed width of 3 cm.
\item The second column (Mutation Type) has a width of 4 cm.
\item The third column (NumSamples) is colored blue.
\item A footnote is added to summarize the content of the table.
}
}
