% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{visualize_cd8_markers}
\alias{visualize_cd8_markers}
\title{Visualize CD8+ T cell Markers}
\usage{
visualize_cd8_markers(
  seurat_obj,
  all_markers,
  cd8_markers,
  reduction_type = "umap",
  color_scheme = c("red", "blue")
)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{all_markers}{A data frame containing all identified markers.}

\item{cd8_markers}{A vector of CD8+ T cell marker genes.}

\item{reduction_type}{Type of dimensionality reduction to use for plotting (default is "umap").}

\item{color_scheme}{A vector of colors for plotting (default is c("red", "blue")).}
}
\value{
Feature plot of CD8+ T cell markers and printed list of enriched clusters.
}
\description{
This function visualizes CD8+ T cell markers using a FeaturePlot and identifies clusters enriched for these markers.
}
