% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{visualize_cd8_subtypes}
\alias{visualize_cd8_subtypes}
\title{Visualize CD8+ Subtypes on UMAP}
\usage{
visualize_cd8_subtypes(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object containing CD8+ T cells with assigned labels.}
}
\value{
UMAP plot of CD8+ subtypes with custom theme.
}
\description{
This function visualizes the CD8+ T cell subtypes on a UMAP plot.
}
