% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{visualize_pca}
\alias{visualize_pca}
\title{Visualize PCA and Elbow Plot}
\usage{
visualize_pca(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA-seq data.}
}
\value{
The Elbow plot, PCA plot, and heatmap of the first 5 principal components.
}
\description{
This function generates an Elbow plot for PCA variance and visualizes the PCA results using a DimPlot.
}
