% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimmunosurv.R
\name{visualize_umap}
\alias{visualize_umap}
\title{Visualize UMAP}
\usage{
visualize_umap(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object containing UMAP results.}
}
\value{
UMAP plot with labels and custom themes.
}
\description{
This function visualizes the UMAP embedding of the Seurat object.
}
